@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.3.1",
-    "react": "^16.10.2",
-    "react-bootstrap": "^1.0.0-beta.14",
-    "react-dom": "^16.10.2",
-    "react-icons": "^3.7.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-icons": "^4.4.0",
+    "react-responsive-carousel": "^3.2.23",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.3",
-    "react-spring": "^8.0.27",
-    "styled-components": "^5.0.1"
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-text-loop": "^2.3.0",
+    "styled-components": "^5.3.5",
+    "typewriter-effect": "^2.18.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -38,7 +38,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^2.1.1"
+    "gh-pages": "^4.0.0"
   },
   "homepage": "https://alenq1.github.io/"
 }
